Revert "Add HDF5 RW interface"
