--- tests/test_suites.h
@@ -1,7 +1,0 @@
-#include <stdlib.h>
-#include <math.h>
-#include <time.h>
-#include "unity.h"
-
-void test_geotransfer(void);
-void test_readHDF5(void);
